+++ packages/tools/request/rollup.config.js
@@ -64,7 +64,7 @@ const browserConfig = {
 		resolve(),
 		commonjs(),
 		removeComments,
-		buble(),
+		buble({objectAssign: 'Object.assign'}),
 	],
 	treeshake
 };