+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 	- renamed `ScreenGauge` to `ScreenSensor`
 	- made `ScreenSensor` into a singleton so that it renders only once
 	- removed `fontSize` property
+	- replaced bound size variables by `ResizeObserver` action
 - add `StorageIO` component
 - add `A11yMenu` component
 - add `FontsLoader` component