+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 - add `A11yMenu` component
 - add `FontsLoader` component
 - replaced some absolute CSS definitions related to text and fonts
+- added missing components to `rollup` input variable
+- fixed linting errors in `src/a11y/menu/settings.js`
 
 ## `@svizzle/atlas` v0.6.0 (next)